feat(repayment): refactor approve loan
