+++ server/controllers/LoanController.js
@@ -7,8 +7,9 @@ import FinAccountRepo from '../repository/FinAccountRepo';
 import LoanTypeRepo from '../repository/LoanTypeRepo';
 import CustomerRepo from '../repository/CustomerRepo';
 import LoanRepo from '../repository/LoanRepo';
+import createLoanRepayments from '../utils/createLoanRepayments';
 
-const { Repayment } = model;
+const { Repayment, sequelize } = model;
 
 /**
  * Controller to handle neccessary loan requests
@@ -115,98 +116,72 @@ class LoanController {
   static async approveLoan(req, res) {
     const { approvedAmount } = req.body;
     const { customerId, loanRefNo } = req.params;
-    // const { id: staffId } = res.locals.user;
-
     const findLoan = () => LoanRepo.getByLoanRefNo(loanRefNo);
     const findCustomer = () => CustomerRepo.getById(customerId);
 
-    const [loan, customer] = await Promise.all([findLoan(), findCustomer()]);
+    try {
+      const [loan, customer] = await Promise.all([findLoan(), findCustomer()]);
 
-    if (!loan) {
-      return response.notFound(res, {
-        message: 'There is no loan with such a loanRefNo',
+      if (!loan) {
+        return response.notFound(res, {
+          message: 'There is no loan with such a loanRefNo',
+        });
+      }
+      if (!customer) {
+        return response.notFound(res, {
+          message: 'There is no customer with such an id',
+        });
+      }
+
+      const {
+        requestAmount,
+        LoanType: { minimumAmount, maximumAmount },
+      } = loan;
+
+      if (Number(approvedAmount) < Number(minimumAmount)) {
+        return response.badRequest(res, {
+          message: `Unable to complete this loan approval, approvedAmount should be equal or greater than the minimum loanable amount of N${minimumAmount}`,
+        });
+      }
+
+      if (Number(approvedAmount) > Number(maximumAmount)) {
+        return response.badRequest(res, {
+          message: `Unable to complete this loan approval, approvedAmount should be equal or less than the maximum loanable amount of N${maximumAmount}`,
+        });
+      }
+
+      if (Number(approvedAmount) > Number(requestAmount)) {
+        return response.badRequest(res, {
+          message: `Unable to complete this loan approval, approvedAmount should be equal or less than the requested amount of N${requestAmount}`,
+        });
+      }
+
+      if (['approved', 'rejected'].includes(loan.approvalStatus)) {
+        return response.badRequest(res, {
+          message: `Unable to complete this loan approval request as it has already been ${
+            loan.approvalStatus
+          }`,
+        });
+      }
+
+      const { repayments, totalPaybackAmount, numberOfPayments } = await createLoanRepayments(
+        loan,
+        approvedAmount,
+      );
+      const message = 'Loan has been approved successfully and the repayment dates have been calculated';
+
+      return response.success(res, {
+        message,
+        loanRefNo,
+        requestAmount,
+        approvedAmount,
+        totalPaybackAmount,
+        numberOfPayments,
+        repayments,
       });
+    } catch (error) {
+      return response.internalError(res, { error });
     }
-    if (!customer) {
-      return response.notFound(res, {
-        message: 'There is no customer with such an id',
-      });
-    }
-
-    const {
-      duration,
-      requestAmount,
-      LoanType: {
-        minimumAmount, maximumAmount, interestRate, payCycle,
-      },
-    } = loan;
-
-    if (Number(approvedAmount) < Number(minimumAmount)) {
-      return response.badRequest(res, {
-        message: `Unable to complete this loan approval, approvedAmount should be equal or greater than the minimum loanable amount of N${minimumAmount}`,
-      });
-    }
-
-    if (Number(approvedAmount) > Number(maximumAmount)) {
-      return response.badRequest(res, {
-        message: `Unable to complete this loan approval, approvedAmount should be equal or less than the maximum loanable amount of N${maximumAmount}`,
-      });
-    }
-
-    if (Number(approvedAmount) > Number(requestAmount)) {
-      return response.badRequest(res, {
-        message: `Unable to complete this loan approval, approvedAmount should be equal or less than the requested amount of N${requestAmount}`,
-      });
-    }
-
-    if (['approved', 'rejected'].includes(loan.approvalStatus)) {
-      return response.badRequest(res, {
-        message: `Unable to complete this loan approval request as it has already been ${
-          loan.approvalStatus
-        }`,
-      });
-    }
-
-    const durationCount = duration[0];
-    const payCycleCount = payCycle === 'Monthly' ? 1 : 4;
-    const descriptionType = payCycle === 'Monthly' ? 'month' : 'week';
-    const numberOfPayments = Number(durationCount) * Number(payCycleCount);
-
-    const interest = approvedAmount * durationCount * (interestRate / 100);
-    const totalPaybackAmount = Number(approvedAmount) + Number(interest);
-
-    const averagePay = totalPaybackAmount / numberOfPayments;
-    const incrementalDays = payCycle === 'Monthly' ? 28 : 7;
-
-    const repayments = [];
-
-    const approvedLoan = await loan.update({ approvedAmount, approvalStatus: 'approved' });
-
-    if (approvedLoan.approvalStatus === 'approved') {
-      range(numberOfPayments).map(number => repayments.push({
-        description: `${descriptionType} ${number + 1}`,
-        dueDate: moment(new Date().getTime() + 86400000 * incrementalDays * (number + 1)).format(
-          'dddd, MMMM DD, YYYY hh:mm:ssA',
-        ),
-        paymentStatus: false,
-        dueAmount: averagePay,
-        loanId: loan.id,
-      }));
-    }
-
-    await Repayment.bulkCreate(repayments);
-
-    const message = 'Loan has been approved successfully and the repayment dates have been calculated';
-
-    return response.success(res, {
-      message,
-      loanRefNo,
-      requestAmount,
-      approvedAmount,
-      totalPaybackAmount,
-      numberOfPayments,
-      repayments,
-    });
   }
 }
